Configured shims package
