+++ src/package.json
@@ -3,6 +3,7 @@
   "type": "module",
   "version": "0.19.1",
   "description": "Emmett - Event Sourcing development made simple",
+  "type": "module",
   "engines": {
     "node": ">=20.11.1"
   },
@@ -86,9 +87,11 @@
     "@types/express": "4.17.21",
     "@types/node": "^22.4.1",
     "@types/supertest": "6.0.2",
+    "web-streams-polyfill": "^4.0.0",
     "supertest": "7.0.0"
   },
   "workspaces": [
+    "packages/emmett-shims",
     "packages/emmett",
     "packages/emmett-postgresql",
     "packages/emmett-esdb",
