+++ src/packages/emmett-shims/tsconfig.eslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.json"
+}
