+++ src/packages/emmett-shims/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.shared.json",
+  "include": ["./src/**/*"],
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "./dist" /* Redirect output structure to the directory. */,
+    "rootDir": "./src"
+  }
+}
