+++ src/packages/emmett-shims/tsup.config.ts
@@ -8,11 +8,12 @@ export default defineConfig({
   dts: true, // generate dts files
   format: ['cjs', 'esm'], // generate cjs and esm files
   minify: true, //env === 'production',
-  bundle: true, //env === 'production',
+  bundle: false, //env === 'production',
   skipNodeModulesBundle: true,
   watch: env === 'development',
   target: 'esnext',
   outDir: 'dist', //env === 'production' ? 'dist' : 'lib',
   entry: ['src/index.ts'],
   sourcemap: true,
+  tsconfig: 'tsconfig.build.json', // workaround for https://github.com/egoist/tsup/issues/571#issuecomment-1760052931
 });
