+++ src/packages/emmett-testcontainers/package.json
@@ -3,6 +3,7 @@
   "version": "0.19.1",
   "type": "module",
   "description": "Emmett - TestContainers - Event Sourcing development made simple",
+  "type": "module",
   "scripts": {
     "build": "tsup",
     "build:ts": "tsc",
