+++ src/packages/emmett-testcontainers/tsup.config.ts
@@ -10,7 +10,6 @@ export default defineConfig({
   minify: true, //env === 'production',
   bundle: true, //env === 'production',
   skipNodeModulesBundle: true,
-  entryPoints: ['src/index.ts'],
   watch: env === 'development',
   target: 'esnext',
   outDir: 'dist', //env === 'production' ? 'dist' : 'lib',
