+++ src/packages/emmett/package.json
@@ -3,6 +3,7 @@
   "type": "module",
   "version": "0.19.1",
   "description": "Emmett - Event Sourcing development made simple",
+  "type": "module",
   "scripts": {
     "build": "tsup",
     "build:ts": "tsc",
@@ -70,6 +71,7 @@
     "emmett": "./dist/cli.js"
   },
   "peerDependencies": {
+    "@event-driven-io/emmett-shims": "0.11.2",
     "@types/async-retry": "^1.4.8",
     "@types/uuid": "^10.0.0",
     "async-retry": "^1.3.3",
