--- src/packages/emmett/src/shims/index.ts
@@ -1,2 +1,0 @@
-export * from './runtime';
-export * from './streams';
