+++ src/packages/emmett/src/streaming/collectors/collect.ts
@@ -1,4 +1,4 @@
-import type { ReadableStream } from 'web-streams-polyfill';
+import type { ReadableStream } from '@event-driven-io/emmett-shims';
 
 export const collect = async <T>(stream: ReadableStream<T>): Promise<T[]> => {
   const results: T[] = [];
