+++ src/packages/emmett/src/streaming/transformations/stopAfter.ts
@@ -1,4 +1,4 @@
-import { TransformStream } from 'web-streams-polyfill';
+import { TransformStream } from '@event-driven-io/emmett-shims';
 
 export const stopAfter = <Item>(stopCondition: (item: Item) => boolean) =>
   new TransformStream<Item, Item>({
