+++ src/packages/emmett/src/streaming/writers/writeToStream.ts
@@ -1,4 +1,7 @@
-import type { ReadableStream, TransformStream } from 'web-streams-polyfill';
+import type {
+  ReadableStream,
+  TransformStream,
+} from '@event-driven-io/emmett-shims';
 import { noMoreWritingOn } from './noMoreWritingOn';
 
 export const writeToStream = async <In, Out>(
