+++ src/packages/emmett/tsup.config.ts
@@ -6,11 +6,10 @@ export default defineConfig({
   splitting: true,
   clean: true, // clean up the dist folder
   dts: true, // generate dts files
-  format: ['cjs', 'esm'], // generate cjs and esm files
+  format: ['esm', 'cjs'], // generate cjs and esm files
   minify: true, //env === 'production',
   bundle: true, //env === 'production',
   skipNodeModulesBundle: true,
-  // entryPoints: ['src/index.ts'],
   watch: env === 'development',
   target: 'esnext',
   outDir: 'dist', //env === 'production' ? 'dist' : 'lib',
