+++ src/tsconfig.json
@@ -6,6 +6,7 @@
   "compilerOptions": {
     "noEmit": true /* Do not emit outputs. */,
     "paths": {
+      "@event-driven-io/emmett-shims": ["./packages/emmett-shims/src"],
       "@event-driven-io/emmett": ["./packages/emmett/src"],
       "@event-driven-io/emmett-expressjs": ["./packages/emmett-expressjs/src"],
       "@event-driven-io/emmett-postgresql": [
